reintegros-backend/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3
PyJWT==2.0.1
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
python-ldap==3.3.1
wheel==0.36.2
pyyaml==5.4.1
uritemplate==3.0.1
requests==2.25.1
django-filter==2.4.0